requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=0.21.0
python-multipart

# Gemini API
google-generativeai>=0.3.0
protobuf>=4.21.0

# LangChain and AI tools
langchain>=0.0.200
langgraph>=0.0.1
crewai>=0.1.0

# Data processing and utilities
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Cybersecurity tools
cryptography>=35.0.0

# HTTP requests (for REST API fallback)
requests>=2.28.0

# Optional: For advanced workflows
transformers>=4.26.0